Ensure that latest last_modified_user_id returned from PUT update
We have an issue since the performance improvements to #299 where the
latest last_modified_user_id wasn't being returned from PUT /update

This was due to the instance of the record not being reloaded after it
was updated given that last_modified_user_id is now a cached lookup
from the versions table.

By specifically reloading this in the controller we get the latest.