Simplify last_modified_user_id by caching it
In order to reduce reading overhead we are now caching the
last_modified_user_id on save for each versioned record.  This allows
the queries to be simpler, and therefore faster, in most cases.

This increases overhead when writing but given the benefits of API
performance when reading it seems worth it.