Upgrade Rails from `7.2.2.1` to `8.0.2` (#438)
Following on from #437, this will bring the Rails version up to `8.0.2`.

All current configuration has been kept, some additional configuration
has been added given the new features of Rails 8.

### Considerations
This upgrade comes with a Deprecation warning to be aware of:

DEPRECATION WARNING: `to_time` will always preserve the full timezone
rather than offset of the receiver in Rails 8.1. To opt in to the new
behavior, set:
 - `config.active_support.to_time_preserves_timezone = :zone`. 

Something to consider for the application when moving to Rails 8.1.

---------

Co-authored-by: Timo Franz <timo@dumpark.com>
Co-authored-by: Philip Arndt <git@p.arndt.io>